backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-debug-toolbar==4.2.0
cryptography==41.0.5
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0 